# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VelozTrade API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /instruments/{symbol}/quote:
    get:
      operationId: getQuote
      tags: [instruments]
      summary: Get real-time quote
      description: Returns current bid/ask price and 24h change for a symbol
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Symbol not found
...
  /watchlist:
    get:
      operationId: getWatchlist
      tags: [watchlist]
      summary: Get watchlist
      description: Returns all symbols in the user watchlist with current quotes
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add symbol to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Symbol added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
        "400":
          description: Symbol already in watchlist
  /watchlist/{symbol}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove symbol from watchlist
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
        "404":
          description: Symbol not in watchlist
...
components:
  schemas:
...
    Quote:
      type: object
      required: [symbol, bid, ask, price, change, changePercent, timestamp]
      properties:
        symbol:
          type: string
        bid:
          type: number
        ask:
          type: number
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        high:
          type: ["number", "null"]
        low:
          type: ["number", "null"]
        volume:
          type: ["number", "null"]
        timestamp:
          type: string
...
    WatchlistItem:
      type: object
      required: [symbol, name, type, price, changePercent, addedAt]
      properties:
        symbol:
          type: string
        name:
          type: string
        type:
          type: string
        price:
          type: ["number", "null"]
        change:
          type: ["number", "null"]
        changePercent:
          type: ["number", "null"]
        bid:
          type: ["number", "null"]
        ask:
          type: ["number", "null"]
        addedAt:
          type: string
    WatchlistInput:
      type: object
      required: [symbol]
      properties:
        symbol:
          type: string
...